+++ src/main/java/com/vetpetmon/wyrmsofnyrus/invasion/HiveCreepSpreadFurther.java
@@ -7,8 +7,8 @@ import com.vetpetmon.wyrmsofnyrus.block.hivecreep.creepStaged;
 import com.vetpetmon.wyrmsofnyrus.block.hivecreep.creepStagedGrass;
 import com.vetpetmon.wyrmsofnyrus.config.Debug;
 import com.vetpetmon.wyrmsofnyrus.config.Invasion;
-import com.vetpetmon.wyrmsofnyrus.synapselib.RNG;
-import com.vetpetmon.wyrmsofnyrus.synapselib.blockUtils;
+import com.vetpetmon.wyrmsofnyrus.synapselib.util.RNG;
+import com.vetpetmon.wyrmsofnyrus.synapselib.util.blockUtils;
 import com.vetpetmon.wyrmsofnyrus.wyrmsofnyrus;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockLog;
