+++ src/main/java/com/vetpetmon/wyrmsofnyrus/invasion/events/massIncursion.java
@@ -1,6 +1,6 @@
 package com.vetpetmon.wyrmsofnyrus.invasion.events;
 
-import com.vetpetmon.wyrmsofnyrus.synapselib.RNG;
+import com.vetpetmon.wyrmsofnyrus.synapselib.util.RNG;
 import net.minecraft.entity.effect.EntityLightningBolt;
 import net.minecraft.world.World;
