+++ src/main/java/com/vetpetmon/wyrmsofnyrus/invasion/events/scoutingPodRaid.java
@@ -3,8 +3,7 @@ package com.vetpetmon.wyrmsofnyrus.invasion.events;
 import com.vetpetmon.wyrmsofnyrus.config.Debug;
 import com.vetpetmon.wyrmsofnyrus.config.Invasion;
 import com.vetpetmon.wyrmsofnyrus.entity.wyrms.EntityCallousPod;
-import com.vetpetmon.wyrmsofnyrus.entity.wyrms.EntityHexePod;
-import com.vetpetmon.wyrmsofnyrus.synapselib.RNG;
+import com.vetpetmon.wyrmsofnyrus.synapselib.util.RNG;
 import com.vetpetmon.wyrmsofnyrus.wyrmsofnyrus;
 import net.minecraft.entity.Entity;
 import net.minecraft.world.World;
