--- src/main/java/com/vetpetmon/wyrmsofnyrus/synapselib/blockUtils.java
@@ -1,9 +1,0 @@
-package com.vetpetmon.wyrmsofnyrus.synapselib;
-
-import net.minecraft.block.material.Material;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-
-public class blockUtils {
-    public static Material getLookingBlockMat(BlockPos pos, World world) {return (world.getBlockState(pos)).getMaterial();}
-}
