+++ src/main/java/com/vetpetmon/wyrmsofnyrus/synapselib/rendering/IHasModel.java
@@ -1,4 +1,4 @@
-package com.vetpetmon.wyrmsofnyrus.synapselib;
+package com.vetpetmon.wyrmsofnyrus.synapselib.rendering;
 
 public interface IHasModel {
     public void registerModels();
